Fix: undo migration of UpcomingEvents model to home app; create view for upcoming events
